Minor update to CommandContainer for allowing beamline actions with input forms based on JSON schema
